Fix exe2minixfs 64 bit integer types to match SWEB types

+++ utils/exe2minixfs/types.h
@@ -23,12 +23,12 @@ typedef uint16_t uint16;
 typedef int32_t int32;
 typedef uint32_t uint32;
 
-typedef uint64_t uint64;
-typedef int64_t int64;
+typedef unsigned long long int uint64;
+typedef long long int int64;
 
 typedef void* pointer;
 
-typedef uint64_t l_off_t;
+typedef uint64 l_off_t;
 
 class FileSystemInfo;
